change env variable setup, move contact components to contact folder
