+++ app/[lang]/(routes)/contact/actions/sendEmail.ts
@@ -2,7 +2,7 @@
 
 import { Resend } from 'resend';
 
-const resend = new Resend(process.env.RESEND_API_KEY);
+const resend = new Resend(process.env.NEXT_PUBLIC_RESEND_API_KEY);
 
 export const sendReservationEmail = () => {
 	resend.emails.send({
