+++ app/[lang]/(routes)/contact/components/RoomReservationForm.tsx
@@ -13,27 +13,37 @@ import {
 import { Input } from '@/app/[lang]/(ui)/components/shadcn/Input';
 import { Textarea } from '@/app/[lang]/(ui)/components/shadcn/Textarea';
 import { DatePicker } from '@/app/[lang]/(ui)/components/shared/DatePicker';
-import { dinnerReservationFormSchema } from '@/app/[lang]/lib/zod/schemas/dinnerReservationFormSchema';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm } from 'react-hook-form';
 import * as z from 'zod';
-import { sendReservationEmail } from '../../actions/sendEmail';
+import { sendReservationEmail } from '../actions/sendEmail';
+import { roomReservationFormSchema } from '@/app/[lang]/lib/zod/schemas/roomReservationFormSchema';
+import {
+	Select,
+	SelectContent,
+	SelectItem,
+	SelectTrigger,
+	SelectValue,
+} from '@/app/[lang]/(ui)/components/shadcn/Select';
 
-export default function DinnerReservationForm() {
-	const form = useForm<z.infer<typeof dinnerReservationFormSchema>>({
-		resolver: zodResolver(dinnerReservationFormSchema),
+export default function RoomReservationForm() {
+	const form = useForm<z.infer<typeof roomReservationFormSchema>>({
+		resolver: zodResolver(roomReservationFormSchema),
 		defaultValues: {
 			firstName: '',
 			lastName: '',
 			email: '',
+			typeOfRoom: '',
+			numberOfNights: 1,
 			numberOfPeople: 1,
+			checkInDate: new Date(),
+			checkOutDate: new Date(),
 			phone: '',
-			reservationDate: new Date(),
 			message: '',
 		},
 	});
 
-	const onSubmit = (values: z.infer<typeof dinnerReservationFormSchema>) => {
+	const onSubmit = (values: z.infer<typeof roomReservationFormSchema>) => {
 		sendReservationEmail();
 		console.log(values);
 	};
@@ -48,7 +58,7 @@ export default function DinnerReservationForm() {
 							name='firstName'
 							render={({ field }) => (
 								<FormItem>
-									<FormLabel>Ime</FormLabel>
+									<FormLabel>First name</FormLabel>
 									<FormControl>
 										<Input type='text' placeholder='First name' {...field} />
 									</FormControl>
@@ -62,7 +72,7 @@ export default function DinnerReservationForm() {
 							name='lastName'
 							render={({ field }) => (
 								<FormItem>
-									<FormLabel>Prezime</FormLabel>
+									<FormLabel>Last name</FormLabel>
 									<FormControl>
 										<Input type='text' placeholder='Last name' {...field} />
 									</FormControl>
@@ -87,10 +97,34 @@ export default function DinnerReservationForm() {
 
 						<FormField
 							control={form.control}
+							name='typeOfRoom'
+							render={({ field }) => (
+								<FormItem>
+									<FormLabel>Type of room</FormLabel>
+									<FormControl>
+										<Select onValueChange={field.onChange} defaultValue={field.value}>
+											<SelectTrigger className='w-full'>
+												<SelectValue placeholder='Type of room' />
+											</SelectTrigger>
+
+											<SelectContent>
+												<SelectItem value='double-bed'>Double bed</SelectItem>
+												<SelectItem value='two-beds'>Two beds</SelectItem>
+												<SelectItem value='single-room'>Single room</SelectItem>
+											</SelectContent>
+										</Select>
+									</FormControl>
+									<FormMessage />
+								</FormItem>
+							)}
+						/>
+
+						<FormField
+							control={form.control}
 							name='phone'
 							render={({ field }) => (
 								<FormItem>
-									<FormLabel>Telefon / mobitel</FormLabel>
+									<FormLabel>Phone</FormLabel>
 									<FormControl>
 										<Input type='tel' placeholder='Phone' {...field} />
 									</FormControl>
@@ -104,7 +138,7 @@ export default function DinnerReservationForm() {
 							name='numberOfPeople'
 							render={({ field }) => (
 								<FormItem>
-									<FormLabel>Broj ljudi</FormLabel>
+									<FormLabel>Number of people</FormLabel>
 									<FormControl>
 										<Input min={1} placeholder='Number of people' {...field} />
 									</FormControl>
@@ -115,10 +149,25 @@ export default function DinnerReservationForm() {
 
 						<FormField
 							control={form.control}
-							name='reservationDate'
+							name='checkInDate'
 							render={({ field }) => (
 								<FormItem>
-									<FormLabel>Datum</FormLabel>
+									<FormLabel>Check in date</FormLabel>
+									<FormControl>
+										<div className='w-full'>
+											<DatePicker />
+										</div>
+									</FormControl>
+								</FormItem>
+							)}
+						/>
+
+						<FormField
+							control={form.control}
+							name='checkOutDate'
+							render={({ field }) => (
+								<FormItem>
+									<FormLabel>Check out date</FormLabel>
 									<FormControl>
 										<div className='w-full'>
 											<DatePicker />
@@ -133,7 +182,7 @@ export default function DinnerReservationForm() {
 							name='message'
 							render={({ field }) => (
 								<FormItem>
-									<FormLabel>Poruka</FormLabel>
+									<FormLabel>Message</FormLabel>
 									<FormControl>
 										<Textarea rows={7} />
 									</FormControl>
@@ -142,7 +191,7 @@ export default function DinnerReservationForm() {
 						/>
 
 						<Button type='submit' className='text-white'>
-							Rezerviraj
+							Make a reservation
 						</Button>
 					</form>
 				</Form>
