+++ app/[lang]/constants/envVariables.ts
@@ -1,1 +1,1 @@
-export const HYGRAPH_CONTENT_API = process.env.HYGRAPH_CONTENT_API;
+export const HYGRAPH_CONTENT_API = process.env.NEXT_PUBLIC_HYGRAPH_CONTENT_API;
