+++ app/[lang]/hooks/useGQLQuery.ts
@@ -11,7 +11,9 @@ export default function useGQLQuery<QueryResultType>(
 	const { data, isLoading, error } = useQuery<QueryResultType>({
 		queryKey: queryKey,
 		queryFn: async () =>
-			request(`${process.env.HYGRAPH_CONTENT_API}`, queryDocument, { locales: [locale] }),
+			request(`${process.env.NEXT_PUBLIC_HYGRAPH_CONTENT_API}`, queryDocument, {
+				locales: [locale],
+			}),
 	});
 
 	return { data, isLoading, error };
